feat: merge consts into config
