+++ crates/common/src/config.rs
@@ -2,6 +2,8 @@
 //!
 //! 这个 crate 中包含了整个系统的配置，需要在多个 crarte 中共享的配置信息。
 
+use sel4::{Word, cap::Endpoint};
+
 /// 服务任务在启动的时候使用的栈的地址。
 pub const SERVICE_BOOT_STACK_TOP: usize = 0x1_0000_0000;
 
@@ -58,3 +60,22 @@ pub const STACK_ALIGN_SIZE: usize = 16;
 
 /// 页共享使用的初始地址
 pub const SHARE_PAGE_START: usize = 0x1_001F_0000;
+
+/// 通过 [sel4::cap::Endpoint] 发送一次数据最大数量
+pub const IPC_DATA_LEN: usize = 120 * 8;
+
+/// 通过 [sel4::cap::Endpoint] 发送数据时，reg 的大小
+pub const REG_LEN: usize = size_of::<Word>();
+
+/// 默认的 Page 用来映射的 Slot
+/// TODO: 找一个更加合适的位置来放置，防止产生冲突
+pub const DEFAULT_PAGE_PLACEHOLDER: u64 = 0;
+
+/// 默认的线程的提示
+pub const DEFAULT_THREAD_NOTIFICATION: u64 = 17;
+
+/// 默认的父进程的 Endpoint
+pub const DEFAULT_PARENT_EP: Endpoint = Endpoint::from_bits(18);
+
+/// 默认的自身提供服务的 Endpoint
+pub const DEFAULT_SERVE_EP: Endpoint = Endpoint::from_bits(19);
