+++ crates/common/src/lib.rs
@@ -4,12 +4,11 @@
 extern crate alloc;
 
 mod obj_allocator;
-mod utils;
 
 pub mod arch;
 pub mod config;
-pub mod consts;
 pub mod ipc_saver;
+pub mod ipcrw;
 pub mod log_impl;
 pub mod page;
 pub mod services;
@@ -17,6 +16,5 @@ pub mod slot;
 
 pub use common_macros::{generate_ipc_send, ipc_trait, ipc_trait_impl};
 pub use obj_allocator::*;
-pub use utils::*;
 
 // FIXME: Make this variable more generic.
