+++ crates/sel4-kit/Cargo.toml
@@ -3,6 +3,9 @@ name = "sel4-kit"
 version = "0.1.0"
 edition = "2024"
 
+[features]
+mcs = []
+
 [dependencies]
 sel4 = { workspace = true }
