+++ services/blk-thread/src/main.rs
@@ -5,14 +5,13 @@ extern crate alloc;
 extern crate blk_thread;
 
 use blk_thread::VIRTIOBLK;
-use common::consts::DEFAULT_SERVE_EP;
+use common::config::DEFAULT_SERVE_EP;
 use sel4::{MessageInfoBuilder, with_ipc_buffer_mut};
 use sel4_runtime::main;
 use srv_gate::blk::BlockIfaceEvent;
 
 #[main]
 fn main() {
-    log::debug!("test test");
     let mut virtio_blk = VIRTIOBLK.lock();
 
     log::debug!("Block device capacity: {:#x}", virtio_blk.capacity());
