+++ services/kernel-thread/src/exception.rs
@@ -4,7 +4,10 @@
 //! 为传统宏内核应用。目前传统宏内核应用的 syscall 需要预处理，将 syscall 指令
 //! 更换为 `0xdeadbeef` 指令，这样在异常处理时可以区分用户异常和系统调用。且不用
 //! 为宏内核支持引入多余的部件。
-use common::{config::PAGE_SIZE, consts::DEFAULT_SERVE_EP, page::PhysPage};
+use common::{
+    config::{DEFAULT_SERVE_EP, PAGE_SIZE},
+    page::PhysPage,
+};
 use sel4::{Fault, UserException, VmFault, cap::Notification, init_thread, with_ipc_buffer};
 use spin::Lazy;
 use syscalls::Errno;
