+++ services/kernel-thread/src/task/init.rs
@@ -1,8 +1,5 @@
 use alloc::{collections::btree_map::BTreeMap, vec::Vec};
-use common::{
-    config::{CNODE_RADIX_BITS, PAGE_SIZE, STACK_ALIGN_SIZE},
-    consts::DEFAULT_PARENT_EP,
-};
+use common::config::{CNODE_RADIX_BITS, DEFAULT_PARENT_EP, PAGE_SIZE, STACK_ALIGN_SIZE};
 use memory_addr::MemoryAddr;
 use sel4::{CNodeCapData, init_thread::slot};
 use sel4_kit::slot_manager::LeafSlot;
