+++ services/lwext4-thread/src/main.rs
@@ -5,8 +5,9 @@ extern crate alloc;
 extern crate lwext4_thread;
 
 use common::{
-    consts::{DEFAULT_SERVE_EP, IPC_DATA_LEN, REG_LEN},
-    services::{IpcBufferRW, root::join_channel},
+    config::{DEFAULT_SERVE_EP, IPC_DATA_LEN, REG_LEN},
+    ipcrw::IpcBufferRW,
+    services::root::join_channel,
 };
 use flatten_objects::FlattenObjects;
 use sel4::{IpcBuffer, MessageInfoBuilder, with_ipc_buffer_mut};
